Correct theme and component resizing behaviour
Components of windows are now notified of theme and bounds related
size changes.

Several major bugs were fixed in SyComponentAbstract:

The wrong delta check was being performed that meant that size changes
were only reported to child components if *both* the x and y delta
were nonzero.

The Math.min() function was used where Math.max() was intended,
which caused components to unconditionally be resized to width 2 and
height 2.

The setBounds() function was not reporting size changes to
child components. This was less of a bug and more unimplemented
functionality.

Fix #3