Move more of the titlebar to components
Titlebars, titlebar texts, and the titlebar buttons are now rendered as
ordinary panels, labels, and buttons with custom themes. The component
API has been modified to allow each component type to return access
to a theme, and the titlebar-specific buttons and labels use this to
return titlebar-specific themes.

Themes have been updated to return functions that order the contents
of titlebars and this has been tested and is correct for the three
provided themes.

The theme designer has been deleted until the new model has been
stabilized.