Implement checkboxes
This splits the original button type into Repeatable and Checkbox
variants, both of which are of type SyButtonType. Both may be
themed separately. This also changes the component-internal
windowTheme() function into one that's now total instead of
raising IllegalStateException if the component is not attached to
a window. Consequently, components are now more likely to do the
correct thing if they have logic that's theme dependent.  This is
true of checkboxes: When pressed, they add a theme-specific image as
a child node to display a checkbox. They cannot do this if they are
not attached to a window, because then there is no theme.

Add a theme listener interface to help implement components that
need to perform actions when the theme changes.

Add more tests for semantics and improve logging efficiency by
adding braces.

Fix #18
Fix #11